requirements.txt
azure-ai-inference==1.0.0b5
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.0
flask-simple-captcha==5.5.5
Faker==33.1.0
Flask-Migrate==4.0.7